+++ snis_packet.h
@@ -144,7 +144,7 @@
 struct update_ship_packet {
 	uint16_t opcode;
 	uint32_t id;
-	uint32_t alive;
+	uint16_t alive;
 	uint32_t x, y, z, vx, vy, vz, yawvel, pitchvel, rollvel;
         uint32_t ntorpedoes;
         uint32_t power;
@@ -195,7 +195,7 @@ struct request_ship_sdata_packet {
 struct update_econ_ship_packet {
 	uint16_t opcode;
 	uint32_t id;
-	uint32_t alive;
+	uint16_t alive;
 	uint32_t x, y, z, vx, vy, vz;
 	union quat orientation;
 	uint8_t shiptype; /* same as ship_sdata_packet subclass */
